+++ test/cypress/support/util.js
@@ -1,53 +1,12 @@
+// General utility functions for the Cypress tests
+
 /**
- *  Goes to lforms_testpage.html
+ *  Goes to addFormToPageTest.html
  */
 export function visitTestPage() {
   cy.visit('test/pages/addFormToPageTest.html');
 }
 
-
-/**
- *  Loads a form from a JSON form definition file from the test/data
- *  directory, and displays the form.  Designed for lforms_testpage.html, but
- *  will work with other pages that have the same "Load" button.
- *  For a general solution using LForms.Util.addFormToPage, see addFormToPage in
- *  this file.
- * @param filepath the path to the form definition file, relative to
- *  test/data/fhirVersion (or just test/data if fhirVersion is not
- *  provided.)
- * @param fhirVersion (optional) the version of FHIR to use.
- */
-/*export function loadFromTestData(filepath, fhirVersion=null) {
-  let testFile = getTestDataPathName(filepath, fhirVersion);
-
-  // Listen for the onFormReadyEvent
-  let formReady = false;
-  cy.get('#formContainer').then((fc)=>fc[0].addEventListener('onFormReady', ()=>formReady=true));
-  cy.get('#formContainer').then((fc)=>fc[0].addEventListener('onError', ()=>formReady=true));
-
-  // Temporarily unhide the file input element.
-  let fileInput = cy.get('#fileAnchor');
-  fileInput.invoke('attr', 'class', '');
-  fileInput.selectFile(testFile);
-  // Re-hide the file input element
-  fileInput.invoke('attr', 'class', 'hide');
-  // wait for the form to render
-  const loadTimeout = 10000;
-  cy.get('.lhc-form-title', {timeout: loadTimeout}).should('be.visible').then(
-    {timeout: loadTimeout}, ()=>{
-
-    return new Cypress.Promise((resolve) => {
-      function checkFormReady() {
-        if (formReady)
-          resolve();
-        else
-          setTimeout(checkFormReady, 50);
-      }
-      setTimeout(checkFormReady, 50);
-    });
-  });
-}
-*/
 
 /**
  *  Calls LForms.Util.addFormToPage
@@ -74,7 +33,7 @@ export function addFormToPage(filePath, container, options) {
 
 
 /**
- *  Returns the full path to a  JSON form definition file in the test/data
+ *  Returns the full path to a JSON form definition file in the test/data
  *  directory.
  * @param filepath the path to the form definition file, relative to
  *  test/data/fhirVersion (or just test/data if fhirVersion is not
@@ -93,13 +52,11 @@ function getTestDataPathName(filepath, fhirVersion=null) {
 
 
 /**
- *  Selects a FHIR version.  Designed for lforms_testpage.html.
+ *  Selects a FHIR version.  Designed for lforms_testpage.html and
+ *  addFormToPageTest.html.
  * @param version the FHIR version to use.
  */
 function setFHIRVersion(version) {
   let fhirVersionField = cy.get('#fhirVersion');
   fhirVersionField.select(version);
 }
-
-
-